requirements.txt
click
diskcache
python-dotenv
openai
prompt_toolkit
tenacity
